+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog
 - [gradle] Cache global `FoundryProperties` instance for subproject use.
 - [gradle] Add `composeRuntimeOnly()` DSL feature.
 - [cli] Check for non-existent paths before traversing in `Path.walkEachFile()`.
+- [gradle] Don't depend on build sources in mod score if `MODULE_SCORE_INCLUDE_GENERATED` is false
 
 0.21.0
 ------